# src/app/api/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: "CommonGroundChecker API"
  version: "1.0.0"
  description: "Machine-readable API description for repository metadata, analysis results, and scoring configuration."
servers:
  - url: "http://localhost:3000"
    description: "Local development server"
paths:
...
  /api/admin/scoring:
    get:
      summary: "Retrieve current scoring configuration"
      responses:
        "200":
          description: "Current scoring configuration"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringConfigResponse"
    post:
      summary: "Save scoring configuration snapshot"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reset:
                  type: boolean
                criterionWeights:
                  type: object
                  additionalProperties:
                    type: number
                criterionRequirementLevels:
                  type: object
                  additionalProperties:
                    type: string
                    enum:
                      - mandatory
                      - recommended
                complexityThreshold:
                  type: number
                complexityMaxCcnThreshold:
                  type: number
                spectralRulesetSource:
                  type: string
      responses:
        "200":
          description: "Saved scoring configuration snapshot"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringConfigResponse"
components:
  schemas:
...
    ScoringConfigResponse:
      type: object
      properties:
        ok:
          type: boolean
        scoringConfigId:
          type:
            - string
            - null
        complexityThreshold:
          type: integer
        complexityMaxCcnThreshold:
          type: integer
        spectralRulesetSource:
          type: string
        criterionWeights:
          type: object
          additionalProperties:
            type: number
        criterionRequirementLevels:
          type: object
          additionalProperties:
            type: string
            enum:
              - mandatory
              - recommended
        defaultCriterionWeights:
          type: object
          additionalProperties:
            type: number
        defaultCriterionRequirementLevels:
          type: object
          additionalProperties:
            type: string
            enum:
              - mandatory
              - recommended
              - informative
        defaultComplexityThreshold:
          type: integer
        defaultComplexityMaxCcnThreshold:
          type: integer
        defaultSpectralRulesetSource:
          type: string
      required:
        - scoringConfigId
        - complexityThreshold
        - complexityMaxCcnThreshold
        - spectralRulesetSource
        - criterionWeights
        - criterionRequirementLevels
        - defaultCriterionWeights
        - defaultCriterionRequirementLevels
        - defaultComplexityThreshold
        - defaultComplexityMaxCcnThreshold
        - defaultSpectralRulesetSource
...